requirements.txt
huggingface_hub
Image
numpy
torch
einops
Pillow
timm
torchvision